refactor: implement suggested changes; extract location input component
